feat: extract verified name status into new file
MST-969
