+++ edx_name_affirmation/api.py
@@ -10,7 +10,8 @@ from edx_name_affirmation.exceptions import (
     VerifiedNameEmptyString,
     VerifiedNameMultipleAttemptIds
 )
-from edx_name_affirmation.models import VerifiedName, VerifiedNameConfig, VerifiedNameStatus
+from edx_name_affirmation.models import VerifiedName, VerifiedNameConfig
+from edx_name_affirmation.statuses import VerifiedNameStatus
 
 log = logging.getLogger(__name__)
